requirements.txt
google-generativeai
google-cloud-aiplatform
torch
opencv-python